+++ Browser/MostPopularNode.m
@@ -82,7 +82,7 @@
 	NSArray				*allStreams			= [[[CollectionManager manager] streamManager] streams];
 	NSArray				*sortedStreams		= [allStreams sortedArrayUsingDescriptors:[NSArray arrayWithObject:descriptor]];
 	NSArray				*filteredStreams	= [sortedStreams filteredArrayUsingPredicate:predicate];	
-	unsigned			count				= (_count > [filteredStreams count] ? [filteredStreams count] : _count);
+	NSUInteger			count				= (_count > [filteredStreams count] ? [filteredStreams count] : _count);
 	
 	[self willChangeValueForKey:@"streams"];
 	[[self streamsArray] replaceObjectsInRange:NSMakeRange(0, [[self streamsArray] count]) withObjectsFromArray:filteredStreams range:NSMakeRange(0, count)];
@@ -98,7 +98,7 @@
 	NSArray				*allStreams			= [[[CollectionManager manager] streamManager] streams];
 	NSArray				*sortedStreams		= [allStreams sortedArrayUsingDescriptors:[NSArray arrayWithObject:descriptor]];
 	NSArray				*filteredStreams	= [sortedStreams filteredArrayUsingPredicate:predicate];
-	unsigned			count				= (_count > [filteredStreams count] ? [filteredStreams count] : _count);
+	NSUInteger			count				= (_count > [filteredStreams count] ? [filteredStreams count] : _count);
 	
 	[self willChangeValueForKey:@"streams"];
 	[[self streamsArray] replaceObjectsInRange:NSMakeRange(0, [[self streamsArray] count]) withObjectsFromArray:filteredStreams range:NSMakeRange(0, count)];
@@ -109,10 +109,10 @@
 
 #pragma mark KVC Mutator Overrides
 
-- (void) insertObject:(AudioStream *)stream inStreamsAtIndex:(unsigned)index
+- (void) insertObject:(AudioStream *)stream inStreamsAtIndex:(NSUInteger)index
 {}
 
-- (void) removeObjectFromStreamsAtIndex:(unsigned)index
+- (void) removeObjectFromStreamsAtIndex:(NSUInteger)index
 {
 	NSAssert([self canRemoveStream], @"Attempt to remove a stream from an immutable MostPopularNode");	
 	AudioStream *stream = [[self streamsArray] objectAtIndex:index];
