+++ Browser/SmartPlaylistNode.m
@@ -133,16 +133,16 @@
 
 #pragma mark KVC Accessor Overrides
 
-- (unsigned)		countOfStreams											{ return [[self smartPlaylist] countOfStreams]; }
-- (AudioStream *)	objectInStreamsAtIndex:(unsigned)index					{ return [[self smartPlaylist] objectInStreamsAtIndex:index]; }
+- (NSUInteger)		countOfStreams											{ return [[self smartPlaylist] countOfStreams]; }
+- (AudioStream *)	objectInStreamsAtIndex:(NSUInteger)index					{ return [[self smartPlaylist] objectInStreamsAtIndex:index]; }
 - (void)			getStreams:(id *)buffer range:(NSRange)aRange			{ return [[self smartPlaylist] getStreams:buffer range:aRange]; }
 
 #pragma mark KVC Mutators Overrides
 
-- (void) insertObject:(AudioStream *)stream inStreamsAtIndex:(unsigned)index
+- (void) insertObject:(AudioStream *)stream inStreamsAtIndex:(NSUInteger)index
 {}
 
-- (void) removeObjectFromStreamsAtIndex:(unsigned)index
+- (void) removeObjectFromStreamsAtIndex:(NSUInteger)index
 {}
 
 @end
